[fits] Remove all unsafe code
Remove all unsafe code by not caching the raw data of the HDU anymore.
The necessity of caching the data is dubious, as we are rarely using it
more than once.
Usually, it is read, then processed to some other data-format that is of
use for the consumer.
When the HDU data is needed, we read it from disk and pass it to the
consumer.

The cache implementation was buggy (probably the cause of some segfaults
on some platforms). If we really need to use cache, we should come up
with a correct implementation at least.